requirements.txt
APScheduler==3.10.4
docker==7.1.0
Flask==3.0.3
getmac==0.9.4
python-dotenv==1.0.1
scapy==2.5.0
